At what point does code refactoring become feng shui?
