+++ viewer/tasks/process.py
@@ -1,10 +1,23 @@
 from background_task import background
-import datetime, pytz, os
 
 from background_task.models import Task
-from viewer.models import Photo, WatchedDirectory
-from viewer.functions.locations import fill_country_cities, fill_location_cities
+from viewer.models import Photo
+from viewer.functions.locations import create_location_events, fill_country_cities, fill_location_cities
 from viewer.functions.location_manager import get_location_manager_report_queue
+
+@background(schedule=0, queue='process')
+def generate_location_events(min_duration=300):
+	"""
+	Creates a set of location Events based on the no-movement periods generated by the Location Manager.
+
+	:param min_duration: Doesn't create Events shorter than this many seconds. Good for avoiding creating 'stopped at a red light' Events.
+	"""
+	if Task.objects.filter(queue='process', task_name__icontains='tasks.generate_location_events').count() > 1:
+		return # If there's already an instance of this task running or queued, don't start another.
+	if len(get_location_manager_report_queue()) > 0:
+		generate_location_events(min_duration=min_duration, schedule=60) # If there are tasks in the location manager, hold back until they finish
+		return
+	create_location_events(min_duration)
 
 @background(schedule=0, queue='process')
 def fill_cities():
